Create param grid (two way factorial)

Create param grid (two way factorial)

create_param_grid_two_way_factorial(param_vals, arm_names, arm_param)

Arguments

param_vals

a named list of length two, giving the values to vary over

arm_names

the names of the arms

arm_param

the parameter to use to assign the arm names

Value

a parameter grid

Examples

param_vals <- list(g_perturbation = seq(0.5, 6, 0.5), pi = seq(0.1, 0.3, 0.1))
arm_names <- c("pi_small", "pi_intermediate", "pi_big")
arm_param <- "pi"
create_param_grid_two_way_factorial(param_vals, arm_names, arm_param)
#>    g_perturbation  pi arm_pi_big arm_pi_intermediate arm_pi_small grid_id
#> 1             0.5 0.1      FALSE               FALSE         TRUE       1
#> 2             1.0 0.1      FALSE               FALSE         TRUE       2
#> 3             1.5 0.1      FALSE               FALSE         TRUE       3
#> 4             2.0 0.1      FALSE               FALSE         TRUE       4
#> 5             2.5 0.1      FALSE               FALSE         TRUE       5
#> 6             3.0 0.1      FALSE               FALSE         TRUE       6
#> 7             3.5 0.1      FALSE               FALSE         TRUE       7
#> 8             4.0 0.1      FALSE               FALSE         TRUE       8
#> 9             4.5 0.1      FALSE               FALSE         TRUE       9
#> 10            5.0 0.1      FALSE               FALSE         TRUE      10
#> 11            5.5 0.1      FALSE               FALSE         TRUE      11
#> 12            6.0 0.1      FALSE               FALSE         TRUE      12
#> 13            0.5 0.2      FALSE                TRUE        FALSE      13
#> 14            1.0 0.2      FALSE                TRUE        FALSE      14
#> 15            1.5 0.2      FALSE                TRUE        FALSE      15
#> 16            2.0 0.2      FALSE                TRUE        FALSE      16
#> 17            2.5 0.2      FALSE                TRUE        FALSE      17
#> 18            3.0 0.2      FALSE                TRUE        FALSE      18
#> 19            3.5 0.2      FALSE                TRUE        FALSE      19
#> 20            4.0 0.2      FALSE                TRUE        FALSE      20
#> 21            4.5 0.2      FALSE                TRUE        FALSE      21
#> 22            5.0 0.2      FALSE                TRUE        FALSE      22
#> 23            5.5 0.2      FALSE                TRUE        FALSE      23
#> 24            6.0 0.2      FALSE                TRUE        FALSE      24
#> 25            0.5 0.3       TRUE               FALSE        FALSE      25
#> 26            1.0 0.3       TRUE               FALSE        FALSE      26
#> 27            1.5 0.3       TRUE               FALSE        FALSE      27
#> 28            2.0 0.3       TRUE               FALSE        FALSE      28
#> 29            2.5 0.3       TRUE               FALSE        FALSE      29
#> 30            3.0 0.3       TRUE               FALSE        FALSE      30
#> 31            3.5 0.3       TRUE               FALSE        FALSE      31
#> 32            4.0 0.3       TRUE               FALSE        FALSE      32
#> 33            4.5 0.3       TRUE               FALSE        FALSE      33
#> 34            5.0 0.3       TRUE               FALSE        FALSE      34
#> 35            5.5 0.3       TRUE               FALSE        FALSE      35
#> 36            6.0 0.3       TRUE               FALSE        FALSE      36